# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MIS Backend HTTP API
  version: "1.0"
  description: |
    Specification of **actual** Micronaut application paths (prefix `/api/...`, not `/api/v1`).
    Complements the generated Swagger (`/swagger-ui`, build artifacts under `META-INF/swagger`).

    ## Authentication
    - **JWT Bearer**: `Authorization: Bearer <token>` (login response includes a `token` field).
    - **Cookie**: HttpOnly `MIS_AUTH=<token>` (set on successful `/api/auth/login` and `/api/auth/first-password-change`).
    - Anonymous: only `/api/auth/login`, `/api/auth/first-password-change`, and `GET`/`POST` under `/api/public/booking/**` (see `application.properties`).

    ## REST access control (`ApiAccessControlFilter`, `/api/**`)
    After authentication, most routes resolve to **exactly one** permission code (`x-required-permission`).
    If the code is absent (`null`), being logged in is enough. If a code is present, it must be in the employee’s effective permission set (role ± overrides via `/api/access/...`).

    Prefix → code mapping (`GET` vs everything else):
    | Prefix / condition | GET | POST, PUT, PATCH, DELETE |
    |---|---|---|
    | `/api/shell/bootstrap?kind=…` | see `kind` below | — |
    | `/api/access` | `permission.read` | `permission.write` |
    | `/api/organizations` | `organization.read` | `organization.write` |
    | `/api/branches` | `branch.read` | `branch.write` |
    | `/api/employees` | `employee.read` | `employee.write` |
...

...

...

    ## Errors
...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in (email + password)
      description: Returns JWT in body and `MIS_AUTH` cookie. For SYSADMIN with `mustChangePassword`, only `passwordChangeRequired` is returned.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: { type: string }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "401": { description: Invalid credentials }
  /api/auth/first-password-change:
    post:
      tags: [Auth]
      summary: First-login password change (SYSADMIN)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password, newPassword]
              properties:
                login: { type: string }
                password: { type: string }
                newPassword: { type: string }
      responses:
        "200": { description: OK, cookie set }
        "401": { description: Authorization error }
        "409": { description: Password change not required }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Clear authentication cookie
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "204": { description: No content }
...
  /api/access/employees/{employeeId}/permissions:
    get:
      tags: [Access]
      summary: Effective permissions for an employee
      x-required-permission: permission.read
      parameters:
        - $ref: "#/components/parameters/employeeId"
      responses:
        "200": { description: OK }
        "404": { description: Employee not found }
    put:
      tags: [Access]
      summary: Update permission overrides
      x-required-permission: permission.write
      parameters:
        - $ref: "#/components/parameters/employeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
        "400": { description: Invalid payload }
  /api/access/employees/{employeeId}/backdate-days:
    get:
      tags: [Access]
      summary: Backdate editing limit (per account)
      x-required-permission: permission.read
      parameters:
        - $ref: "#/components/parameters/employeeId"
      responses:
        "200": { description: OK }
    put:
      tags: [Access]
      summary: Set backdate limit
      x-required-permission: permission.write
      parameters:
        - $ref: "#/components/parameters/employeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days: { type: integer, minimum: 0 }
      responses:
        "200": { description: OK }
...
  /api/employees:
    get:
      tags: [Employees]
      summary: List employees
      x-required-permission: employee.read
      parameters:
        - $ref: "#/components/parameters/pageablePage"
        - $ref: "#/components/parameters/pageableSize"
      responses:
        "200": { description: Page }
    post:
      tags: [Employees]
      summary: Create employee
      x-required-permission: employee.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
...
  /api/time-slots:
    get:
      tags: [TimeSlots]
      summary: List time slots
      x-required-permission: schedule.view
      parameters:
        - $ref: "#/components/parameters/pageablePage"
        - $ref: "#/components/parameters/pageableSize"
      responses:
        "200": { description: Page }
    post:
      tags: [TimeSlots]
      summary: Create time slot
      x-required-permission: schedule.view
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
  /api/time-slots/for-online-booking:
    get:
      tags: [TimeSlots]
      summary: Time slots for online booking (one day)
      x-required-permission: schedule.view
      parameters:
        - name: branchId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: employeeId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: slotDate
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /api/time-slots/{id}:
    get:
      tags: [TimeSlots]
      summary: Time slot by id
      x-required-permission: schedule.view
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "200": { description: OK }
    patch:
      tags: [TimeSlots]
      summary: Update time slot
      x-required-permission: schedule.view
      parameters:
        - $ref: "#/components/parameters/idUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
    delete:
      tags: [TimeSlots]
      summary: Delete time slot
      x-required-permission: schedule.view
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from `token` field in `/api/auth/login` response
    cookieAuth:
      type: apiKey
      in: cookie
      name: MIS_AUTH
      description: HttpOnly cookie (alternative to Bearer)
  parameters:
    idUuid:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    employeeId:
      name: employeeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    pageablePage:
      name: page
      in: query
      schema: { type: integer, minimum: 0 }
      description: Page number (Micronaut Pageable)
    pageableSize:
      name: size
      in: query
      schema: { type: integer, minimum: 1 }
      description: Page size
...